Refactor Event command into CRUD
