+++ src/main/java/seedu/address/logic/commands/create/AddEventCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands.event;
+package seedu.address.logic.commands.create;
 
 import static seedu.address.logic.parser.CliSyntax.EVENT_COMMAND_WORD;
 import static seedu.address.logic.parser.event.EventCliSyntax.PREFIX_EVENT_END_LONG;
@@ -7,7 +7,6 @@ import static seedu.address.logic.parser.event.EventCliSyntax.PREFIX_EVENT_NAME_
 import static seedu.address.logic.parser.event.EventCliSyntax.PREFIX_EVENT_START_LONG;
 
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.create.AddCommand;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
