+++ src/main/java/seedu/address/logic/commands/delete/ClearEventCommand.java
@@ -1,8 +1,7 @@
-package seedu.address.logic.commands.event;
+package seedu.address.logic.commands.delete;
 
 import static seedu.address.logic.parser.CliSyntax.EVENT_COMMAND_WORD;
 
-import seedu.address.logic.commands.delete.ClearCommand;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
 import seedu.address.model.event.EventManager;
