+++ src/main/java/seedu/address/logic/commands/delete/DeleteEventCommand.java
@@ -1,10 +1,9 @@
-package seedu.address.logic.commands.event;
+package seedu.address.logic.commands.delete;
 
 import static seedu.address.logic.parser.CliSyntax.EVENT_COMMAND_WORD;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.delete.DeleteCommand;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
