+++ src/main/java/seedu/address/logic/commands/read/DisplayEventInformationCommand.java
@@ -1,10 +1,9 @@
-package seedu.address.logic.commands.event;
+package seedu.address.logic.commands.read;
 
 import static seedu.address.logic.parser.CliSyntax.EVENT_COMMAND_WORD;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
-import seedu.address.logic.commands.read.DisplayInformationCommand;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
