+++ src/main/java/seedu/address/logic/parser/HelpCommandParser.java
@@ -20,18 +20,18 @@ import seedu.address.logic.commands.read.FilterContactCommand;
 import seedu.address.logic.commands.read.InfoContactCommand;
 import seedu.address.logic.commands.read.ListContactCommand;
 import seedu.address.logic.commands.update.RemoveTagFromContactCommand;
-import seedu.address.logic.commands.event.AddContactToEventCommand;
-import seedu.address.logic.commands.event.AddContactToLogEventCommand;
-import seedu.address.logic.commands.event.AddEventCommand;
-import seedu.address.logic.commands.event.AddTagToEventCommand;
-import seedu.address.logic.commands.event.ClearEventCommand;
-import seedu.address.logic.commands.event.DeleteEventCommand;
-import seedu.address.logic.commands.event.DisplayEventInformationCommand;
-import seedu.address.logic.commands.event.FilterEventCommand;
-import seedu.address.logic.commands.event.ListEventCommand;
-import seedu.address.logic.commands.event.RemoveContactFromEventCommand;
-import seedu.address.logic.commands.event.RemoveContactFromLogEventCommand;
-import seedu.address.logic.commands.event.RemoveTagFromEventCommand;
+import seedu.address.logic.commands.update.AddContactToEventCommand;
+import seedu.address.logic.commands.update.AddContactToLogEventCommand;
+import seedu.address.logic.commands.create.AddEventCommand;
+import seedu.address.logic.commands.update.AddTagToEventCommand;
+import seedu.address.logic.commands.delete.ClearEventCommand;
+import seedu.address.logic.commands.delete.DeleteEventCommand;
+import seedu.address.logic.commands.read.DisplayEventInformationCommand;
+import seedu.address.logic.commands.read.FilterEventCommand;
+import seedu.address.logic.commands.read.ListEventCommand;
+import seedu.address.logic.commands.update.RemoveContactFromEventCommand;
+import seedu.address.logic.commands.update.RemoveContactFromLogEventCommand;
+import seedu.address.logic.commands.update.RemoveTagFromEventCommand;
 import seedu.address.logic.commands.todo.AddContactToTodoCommand;
 import seedu.address.logic.commands.todo.AddTagToTodoCommand;
 import seedu.address.logic.commands.todo.AddTodoCommand;
