+++ src/main/java/seedu/address/logic/parser/event/DeleteEventCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser.event;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.event.DeleteEventCommand;
+import seedu.address.logic.commands.delete.DeleteEventCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
