+++ src/main/java/seedu/address/logic/parser/event/FilterEventCommandParser.java
@@ -14,8 +14,8 @@ import java.util.List;
 import seedu.address.commons.core.Operator;
 import seedu.address.commons.core.Pair;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.event.FilterEventCommand;
-import seedu.address.logic.commands.event.FilterEventCommand.EventPredicate;
+import seedu.address.logic.commands.read.FilterEventCommand;
+import seedu.address.logic.commands.read.FilterEventCommand.EventPredicate;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
